+++ akshare/stock/stock_info_em.py
@@ -1,7 +1,7 @@
 # -*- coding:utf-8 -*-
 # !/usr/bin/env python
 """
-Date: 2022/1/7 16:21
+Date: 2022/4/19 20:21
 Desc: 东方财富网-个股-股票信息
 http://quote.eastmoney.com/concept/sh603777.html?from=classic
 """