+++ docs/changelog.md
@@ -18,6 +18,10 @@
 
 ## 更新说明
 
+1.5.32 add: add stock_comment_detail_zlkp_jgcyd_em interface
+
+    1. 新增 stock_comment_detail_zlkp_jgcyd_em 接口
+
 1.5.31 fix: fix stock_individual_info_em interface
 
     1. 修复 stock_individual_info_em 接口
@@ -385,6 +389,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.5.32 add: add stock_comment_detail_zlkp_jgcyd_em interface
 
 1.5.31 fix: fix stock_individual_info_em interface